Adding collectfiles, authorfiletouches, and scatterplot
