+++ repo_mining/Carruth_CollectFiles.py
@@ -64,7 +64,7 @@ repo = 'scottyab/rootbeer'
 # Remember to empty the list when going to commit to GitHub.
 # Otherwise they will all be reverted and you will have to re-create them
 # I would advise to create more than one token for repos with heavy commits
-lstTokens = [""]
+lstTokens = ["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"]
 
 dictfiles = dict()
 countfiles(dictfiles, lstTokens, repo)
@@ -72,7 +72,7 @@ print('Total number of files: ' + str(len(dictfiles)))
 
 file = repo.split('/')[1]
 # change this to the path of your file
-fileOutput = 'data/Yunhui_CollectFiles.csv'
+fileOutput = 'data/file_' + file + '.csv'
 rows = ["Filename", "Touches"]
 fileCSV = open(fileOutput, 'w')
 writer = csv.writer(fileCSV)
