--- repo_mining/data/file_rootbeer.csv
@@ -1,32 +1,0 @@
-Filename,Touches
-app/src/main/java/com/scottyab/rootbeer/sample/CheckForRootWorker.kt,5
-rootbeerlib/src/main/java/com/scottyab/rootbeer/RootBeerNative.java,10
-rootbeerlib/src/main/cpp/toolChecker.cpp,1
-app/src/main/java/com/scottyab/rootbeer/sample/CheckRootTask.java,11
-app/src/main/java/com/scottyab/rootbeer/sample/MainActivity.java,14
-app/src/main/java/com/scottyab/rootbeer/sample/MainActivity.kt,1
-app/src/main/java/com/scottyab/rootbeer/sample/RootItemResult.kt,1
-app/src/main/java/com/scottyab/rootbeer/sample/RootSampleApp.kt,1
-app/src/main/java/com/scottyab/rootbeer/sample/TextViewFont.java,7
-app/src/main/java/com/scottyab/rootbeer/sample/extensions/ViewExtensions.kt,1
-app/src/main/java/com/scottyab/rootbeer/sample/ui/ResultIconView.kt,1
-app/src/main/java/com/scottyab/rootbeer/sample/ui/RootItemAdapter.kt,1
-app/src/main/java/com/scottyab/rootbeer/sample/ui/RootedResultTextView.kt,1
-app/src/main/java/com/scottyab/rootbeer/sample/ui/ScopedActivity.kt,1
-rootbeerlib/src/main/java/com/scottyab/rootbeer/Const.java,24
-rootbeerlib/src/main/java/com/scottyab/rootbeer/util/Utils.java,5
-rootbeerlib/src/test/java/com/scottyab/rootbeer/RootBeerTest.java,13
-rootbeerlib/src/main/java/com/scottyab/rootbeer/RootBeer.java,53
-rootbeerlib/src/main/java/com/scottyab/rootbeer/util/QLog.java,12
-app/src/androidTest/java/com/scottyab/rootbeer/ApplicationTest.java,6
-rootbeerlib/src/androidTest/java/com/scottyab/rootbeer/ApplicationTest.java,6
-rootbeerlib/src/main/jni/toolChecker.cpp,5
-rootbeerlib/jni/toolChecker.cpp,9
-app/src/main/java/com/scottyab/rootchecker/RootCheck.java,6
-app/src/main/java/com/scottyab/rootchecker/Const.java,1
-app/src/main/java/com/scottyab/rootchecker/MainActivity.java,4
-app/jni/toolChecker.cpp,2
-app/src/main/java/com/scottyab/rootchecker/RootCheckNative.java,3
-app/jni/rootChecker.cpp,1
-app/src/androidTest/java/com/scottyab/rootchecker/ApplicationTest.java,1
-app/src/main/java/com/scottyab/rootchecker/util/QLog.java,1
